requirements.txt
# Database - using newer versions compatible with Python 3.13
asyncpg>=0.29.0
pgvector>=0.2.4

# CLI
click>=8.1.7
rich>=13.7.0

# PDF processing
PyPDF2>=3.0.1
pdfplumber>=0.10.3

# HTTP requests
httpx>=0.26.0
aiohttp>=3.9.1

# Data validation - using newer version for Python 3.13 compatibility
pydantic>=2.6.0

# Environment
python-dotenv>=1.0.0

# Logging
structlog>=23.2.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# YAML processing
PyYAML>=6.0.1

# OpenAI API (optional)
openai>=1.12.0

# Development
black>=23.12.1
isort>=5.13.2
mypy>=1.8.0